refactor(all): splitted comps to each files
