+++ src/pages/Chaejun/BoardContext.tsx
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import { BoardContextInterface } from "./utils/board";
+
+export const BoardContext = createContext<BoardContextInterface | null>(null);
