+++ src/pages/Chaejun/components/ResetButton.tsx
@@ -1,0 +1,19 @@
+import { useContext } from "react";
+import { BoardContext } from "../BoardContext";
+import { initializeBoard } from "../utils/board";
+import { BoardContextInterface } from "../utils/board";
+
+export function ResetButton() {
+  const { board, setBoard } = {
+    ...useContext(BoardContext),
+  } as BoardContextInterface;
+  return (
+    <button
+      type="button"
+      className="grow rounded-md border-2 border-black"
+      onClick={() => initializeBoard(setBoard, board.rows, board.columns)}
+    >
+      Reset
+    </button>
+  );
+}
