+++ src/pages/Chaejun/index.tsx
@@ -1,0 +1,59 @@
+import { useState } from "react";
+import MainLayout from "../../layouts/MainLayout";
+import { AlgoButton } from "./components/AlgoButton";
+import { BoardContext } from "./BoardContext";
+import { NumberInput } from "./components/NumberInput";
+import { ResetButton } from "./components/ResetButton";
+import { SetMazeButton } from "./components/SetMazeButton";
+import { Table } from "./components/Table";
+import {
+  initializeBoard,
+  create2DArray,
+  safeInput,
+  BoardInterface,
+} from "./utils/board";
+import { BFS, DFS } from "./utils/algorithm";
+
+const Chaejun = () => {
+  const [board, setBoard] = useState<BoardInterface>({
+    rows: 10,
+    columns: 10,
+    maze: create2DArray(10, 10),
+    currentPoint: { x: 0, y: 0 },
+    deque: [{ x: 0, y: 0 }],
+  });
+
+  return (
+    <MainLayout>
+      <BoardContext.Provider value={{ board, setBoard }}>
+        <h1 className="text-4xl font-bold">DFS/BFS</h1>
+        <form className="flex flex-col gap-2 py-4">
+          <div className="grid grid-cols-2 gap-2">
+            <NumberInput
+              props={{ label: "columns", board, setBoard }}
+              onChange={(e) =>
+                initializeBoard(setBoard, board.rows, safeInput(e))
+              }
+            />
+            <NumberInput
+              props={{ label: "rows", board, setBoard }}
+              onChange={(e) =>
+                initializeBoard(setBoard, safeInput(e), board.columns)
+              }
+            />
+            <ResetButton />
+            <SetMazeButton />
+            <AlgoButton callbackFn={DFS} label={"DFS"} />
+            <AlgoButton callbackFn={BFS} label={"BFS"} />
+          </div>
+        </form>
+        <h2 className="text-2xl font-bold">Board</h2>
+        <div className="block rounded-md border-2 border-black p-3">
+          <Table />
+        </div>
+      </BoardContext.Provider>
+    </MainLayout>
+  );
+};
+
+export default Chaejun;
